Refactored Optic and Sling service classes.

+++ src/main/java/com/awoodcock/armory/service/accessories/SlingService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,8 +29,9 @@ public class SlingService {
     }
 
     // READ
-    public Optional<Sling> getSlingById(int id) {
-        return slingRepository.findById(id);
+    public Sling getSlingById(int id) {
+        Optional<Sling> sling = slingRepository.findById(id);
+        return sling.orElse(null);
     }
 
     public List<Sling> getAllSlings() {
@@ -37,10 +39,53 @@ public class SlingService {
     }
 
     // UPDATE
+    public Sling updateSlingById(int id, Sling sling) {
+        Optional<Sling> slingOptional = slingRepository.findById(id);
 
+        if (slingOptional.isPresent()) {
+            Sling originalSling = slingOptional.get();
+
+            // name
+            if (Objects.nonNull(sling.getName()) && !"".equalsIgnoreCase(sling.getName())) {
+                originalSling.setName(sling.getName());
+            }
+
+            // description
+            if (Objects.nonNull(sling.getDescription()) && !"".equalsIgnoreCase(sling.getName())) {
+                originalSling.setDescription(sling.getDescription());
+            }
+
+            // price
+            if (Objects.nonNull(sling.getPrice()) && sling.getPrice() != 0) {
+                originalSling.setPrice(sling.getPrice());
+            }
+
+            // weight
+            if (Objects.nonNull(sling.getWeight()) && sling.getWeight() != 0) {
+                originalSling.setWeight(sling.getWeight());
+            }
+
+            // numPoints
+            if (Objects.nonNull(sling.getNumPoints()) && sling.getNumPoints() != 0) {
+                originalSling.setNumPoints((sling.getNumPoints()));
+            }
+
+            // attachmentMethod
+            if (Objects.nonNull(sling.getDescription()) && !"".equalsIgnoreCase(sling.getName())) {
+                originalSling.setDescription(sling.getDescription());
+            }
+
+            return slingRepository.save(originalSling);
+        }
+        return null;
+    }
     // DELETE
-    public void deleteSling(int id) {
-        slingRepository.deleteById(id);
+    public String deleteSling(int id) {
+        if (slingRepository.findById(id).isPresent()) {
+            slingRepository.deleteById(id);
+            return "Sling successfully deleted.";
+        }
+        return "Sling not found within database.";
     }
 
 }
